Lower MM MTU from 1458 to 1380
We were using an MTU of 1458 for our GRE tunnels for MM tests.
1500 (default eth MTU)
- 20 Bytes IPv4 header
-  4 Bytes GRE header
- 14 Bytes Ethernet header whithin GRE tunnel
-  4 Bytes Ethernet VLAN TAG header whithin GRE tunnel
=> 1458 Bytes

With some MM workers in PRG2 and NUE2 we need to take
the path MTU between our locations into account,
which is 1422 (78 Bytes overhead).

So we have a usable MTU of 1380 left.

Ticket: https://progress.opensuse.org/issues/152389