rsync_server: ensure /etc/rsyncd.secrets has mode 600
In the past, the package installed an /etc/rsyncd.secrets file with mode 0600.
The test then added credentials there using echo user:pass > /etc/rsyncd.secrets
(preserving the mode of the file)

The package changed to install template config to /usr/etc. As the file now does
not exist in /etc, echo creates a new file with mode 644 - which in turn is
denied by rsync's default config (strict modes on)