Wait until prometheus node_exporter is ready before testing
The sles4sap/monitoring_services test module installs and start the
prometheus node_exporter service and checks that it is running by
downloading the metrics with curl. In some scenarios, enabling and
starting the service and trying to curl the metrics right away leads to
failures, as the node_exporter is still initializing. This commit adds a
30s loop to check the service is ready and listening.