containers: only run man if the actual command is present
Newer versions of microos-tools install an alias man=man-online.
The 'command -v man' check in the interactive shell follows the alias
and reports it as known command. Subsequently, the test calls
bash -eox pipefail script.sh though, in which case script.sh does not
have access to the aliased commands. As a consequence, at this point,
'man' fails to be executed.