Add installation and smoke test for SAP ASE
Add test modules to install SAP ASE and to perform some smoke tests on
the installed software: check that the server is running, query it to
get the installed SAP ASE version, stop and start the database server,
query the cockpit service status and stop and start the cockpit.

Included with the test modules is a YAML schedule loading the new
modules, and also the changes required for SAP ASE in `lib/sles4sap.pm`.

Also included in the commit is the change from the deprecated
`saptune daemon` commands to the new `saptune service` commands, and
minor typo fixes and formatting to the POD documentation in
`lib/sles4sap.pm`.