Increase wait time for cs_wait_for_idle in CSP
On prem tests wait iby default for cs_wait_for_idle to finish in 120
seconds, which can be scaled using the TIMEOUT_SCALE setting. Similar
tests in cloud service providers wait by default for 60 seconds, and
we have observed some tests failing in this step. This commit increases
the timeout to 120 seconds.