Rework ENSA2 test to avoid race conditions
Related Ticket: TEAM-9169

The test, as is, uses the supportserver job like other MM scenarios but also
inserts an extra module to set up an NFS server on the supportserver. This is
not using the nfs role of the supportserver but a separate module that is
inheriting from the sles4sap base. Adding the module exposes a race condition
where the nodes reach BARRIER_HA_ensa before it is created by the supportserver.

To fix this we split barrier_init.pm into two modules.
1. ensa_barriers.pm creates the barriers that are necessary for the scenario.
This includes the barrier that supportserver is expected to hit as well as the
node-specific barriers, but does not contain any others.
2. ensa_finish_supportserver.pm uploads the iscsi info that the nodes need
and signals to them that the supportserver has completed the setup, much like
barrier_init.pm originally did.

Unrelated to fixing the race condition:
1. The part of ensa_filesystems.pm that had to do with setting up the NFS server
and clients was extracted to netweaver_nfs_setup.pm. This makes the NFS setup
possibly reusable for other netweaver tests and reduces the extra code that is
in the supportserver scenario
2. The modules that are specific to this scenario have been moved to a newly
created 'ensa' directory
3. Two library functions where created to lib/hacluster.pm. get_cluster_info()
parses the CLUSTER_INFOS variable. It assumes that there is only one cluster
with many nodes. generate_lun_list() generates the file that the nodes scp to
know how to use iscsi to communicate

Because of the reordering, the schedule was completely split from the other ha
jobs into an ensa directory and yaml