Update lib/hacluster POD documentation
Documentation of lib/hacluster was missing for some of the functions and
for some exported variables. This commits adds it, while also fixing
some minor issues such as the formatting of the input arguments to some
of the functions and the font size on the documentation of the newer functions.

Michele's