Split barrier_init and ensa_filesystems
barrier_init is too verbose, doesn't only initialize barriers and
prevents us from having an extra configuration step for the
supportworker. Additionaly it can be prone to race conditions. To
avoid that I am splitting it into a barrier creation module and
module that will signal to the nodes that the supportserver
configuration is done.

ensa_filesystems is split into ensa_filesystems and netweaver_nfs_setup
so that it has no code that needs to run on the supportserver. In
addition, it allows the nfs_setup module to be used by other scenarios
in the future.