Run OpenStack/Cloud image in QEMU
In order to test cloud-init and Cloud image outside of OpenStack we need
to adapt our testing to run the image in QEMU.

In OpenStack tests, we can generate cloud config in the runtime, hence
we can add ssh keys or SCC details. For QEMU, the cloud config is stored
in a qcow2 image as it is in Combustion or Ignition cases and therefore
the configuration is static.

- ticket: https://progress.opensuse.org/issues/156151