Add needle check for successful iPXE start before changing boot device
When we boot slow systems, a hard-coded timeout is either too long or
too short, depending on the hardware used.
Instead of guessing, we should add a needle check to see when iPXE has
successfully started to load a kernel image, then we can safely change
UEFI boot device again.