Move bci_prepare to "Create bootable host HDDs "
The bci_prepare step is an installation step that is executed for every BCI test run. We could move this to the host generation, and void running this for each test run.Small modifications in bci_prepare might be necessary, e.g. that we install all container engines at all times.