Update the system first before installing incidents in SLE Micro
We are booting GA images for maintenance updates. Ideally, we should
install the incidents under tests on an updated system.
It has been observed that installing everything at once can cause
false positives and packages not being updated.

With this approach, we first update the system and after reboot, we
add and install the incidents using the patch option.