Adjust the saptune test on power9
We run saptune 'notes' and 'solutions' on power9 lpar. So we need a add
a yaml file to define the test modules.
There are some changes on saptune test, such as on power platform, the
default value of `transparent_hugepage` is changed to 'madvise'. So we
need to adjust our code to check that.

Related: https://jira.suse.com/browse/TEAM-9086