Use podman instead of docker in hawk_gui test module
The `ha/hawk_gui` test module uses a python+selenium test packaged in a
docker/podman container to ensure all required dependencies are
installed. Early versions of the module would install a fixed docker
version downloaded from download.docker.com, which will then be used to
download the docker image and run the container, but starting on 15-SP6,
this old version of docker clashes with the cgroups configuration of the
base 15-SP6 system, preventing the container from running. This commit
switches the docker commands for podman, as the latter is supported by
SUSE and we can always get the latest version using `zypper in`. These
changes also mean that HAWK client tests need to run from a system
where podman can be installed (i.e., repositories where podman package
is available are already added to the testing system.)