Skip version check in Tumbleweed when BUILD setting is non-numeric
The test module `console/check_os_release` is used to verify the
contents of `/etc/os-release` in Tumbleweed, Leap and SLES (and SLES for
SAP Applications). Since Tumbleweed is a rolling release, instead of a
version number, `/etc/os-release` has a date, which the test module
gets from the BUILD setting to check for correctness. However, during
verification runs, usually the value of the BUILD setting is
replaced for something different, which can cause a verification run
from an unrelated commit to fail on `console/check_os_release`. This
commit checks that BUILD is a number and if it's not, it will skip the
VERSION, VERSION_ID and CPE_NAME check in Tumbleweed. Leap and SLES
code remains the same.