ipxe_install: Make the set_bootscript_hdd() delay configurable (#17783)
* ipxe_install: Make the set_bootscript_hdd() delay configurable

Some workers need more than 2 minutes to reach PXE bootloader after
power-on. Add a variable PXE_BOOT_TIME to control the bootscript
switch delay for UEFI machines so that it can be configured
on each worker.

* ipxe_install: Change condition for HDD bootscript workaround

The HDD bootscript workaround is needed only for some machines.
It should be enabled on a per-worker basis or in specific jobs where
YaST should not modify default boot order.