Add support to publish qcow2 images on HA cluster schedules
This introduces into the basic HA cluster schedules:

- `schedule/ha/bv/basic_cluster_node.yaml`
- `schedule/ha/bv/ha_supportserver.yaml`

Support to use them to publish the qcow2 images after the cluster jobs
finish, by adding the CREATE_HDD=1 setting in the jobs.

Also included in the commit is the a new module for the support server
to clear the lease database from the DHCP server, as this would allow
the migration jobs which run from the generated qcow2 to get the same IP
addresses as the ones using during qcow2 creation without needing to fix
the macaddresses of the SUTs.