Remove openssh-server-config-disallow-rootlogin before a reboot in s390x
When testing any patch that rebuilds openssh, openqa fails only when
testing in s390x (like in https://openqa.suse.de/tests/14804700).
This happens because the test installs all packages that are part of
the patch (which includes openssh-server-config-disallow-rootlogin,
even if that package is never installed automatically) to test that
the system can reboot successfully afterwards. The problem is that
in s390x, after a reboot, openqa needs to ssh to the system to
reconnect the tty, which now fails because that package disallows
root to do a ssh password login.

Make sure that openssh-server-config-disallow-rootlogin is not
installed since in s390 we need to ssh to the system to reconnect
to the tty after a reboot.