Tiny fixes for sle-micro container updates testing
patterns-microos-fips is being used for sle-micro from 5.1 to 5.5.
Fips setup tests should look for this pattern instead of plain fips

As a technical debt caused by the split of responsibilities of sle-micro
testing. We have a wrong schedule of modules. firewalld is not present
in older sle-micro deployments done by the installator