Update nfs_baremetal.yaml to include system installation
For baremetal jobs, we need to create barriers even before we start the
system installation, as times needed for installation and reboot vary
drastically, and more often than not tests fail because the wrong job
is faster and fails with "lock owner already finished".
Having a dedicated schedule will fix this issue by creating barriers even
before we start with system boot.