Improve robustness of cloud-init wait (#19754)
The API to wait the end of the cloud-init execution has an internal
(within Azure) and external (in openQA testapi) timeout. The external
one is larger to give some time to the az cli to terminate smoothly when
the internal timeout elapses. Increase the delta from 10sec to 1 minute.
Add timeout to other API too.
Improve diagnostic log upload code.