Make nfs tests more configurable and support for bonding interfaces
This allows the NFS tests to be controlled by a bunch of (optional)
variables. Also this checks for kernel support for NFSv3 and NFSv4
before doing anything related, although normally we expect this to be
present on the kernels.
We also check for NFSv4.1 and NFSv4.2 support, even this is unused for
now, we need this later for enabling multipath testing.

Also support link aggregation (bonding):
If using a switch with bonding, make sure to have the switch ports
configured in a compatible way of the bonding interfaces. For a direct
connection, nothing more is to be done here.

We need to specify some more variables in order to enable and use this
test:
- NFS_BOND
  only when enabled by setting to '1' we configure the bond

- NFS_IF_1 and NFS_IF_2
  specfify which ethernet devices should be configured for use with
  the bond interface

- NFS_BOND_IP
  The IP address configuration used for bond0. Specify like
  192.168.100.10/24

- BOND_MODE (optional)
  Allow to configure the bond mode to use, defaults to 802.3ad

Also make sure to set SERVER_NODE and CLIENT_NODE to these IPs