Add wait_for_idle_cluster after resource move
The `ha/filesystem` module does a `crm resource migrate` when testing
drbd_passive, moving the resource from one node to the other. In some
slower architectures, the resource has not finished starting in the
target node before testing which can cause failures. This commit works
around the issue by calling on `cs_wait_for_cluster` right after the
move operation to ensure the cluster is ready for the test. Also part of
the commit is the change of `crm resource migrate` to `crm resource
move` as the migrate command is deprecated in favor of `move`.