Introduce HPVS testing in IBM Cloud
[IBM Hyper Protect Platform](https://www.ibm.com/downloads/cas/GPVMWPM3)
is another "host system" where we need to run our containers.
The specific is that there is no access on the VM itself, container is
configure through `compose` and `workload contract` yamls.

VPC, network and floating IP segment is as of now configure manually,
but should be moved into qe-c infra repository.

Other that this test can be considered as an IBMCloud CLI smoke test as
well

- ticket: [[IBM-HPVS] Establish workflow to run custom BCI container
  commands and obtain the logs](https://progress.opensuse.org/issues/163190)