Change regex to search for devices
Previously, we explicitely searched for "infiniband" devices from
Mellanox.
However, if we have those configured as Ethernet devices, we won't
find any devices. We know we only have ConnectX-5 devices for
InfiniBand (but on the same machines we have ConnectX-4 LX which only
do Ethernet - so let's explicitely search for ConnectX-5 devices instead.