Check SMELT to decide when SLES+HA tests can be skipped on QAM
The `lib/hacluster::is_not_maintenance` function was originally created
to skip resource-expensive test modules when the package under test in
Maintenance is not related to these tests (for example, `samba` or
`ctdb` for `ha/ctdb` and `ha/filesystem`, `haproxy` for `ha/haproxy` or
`drbd` for `ha/drbd_passive` and `ha/filesystem`). The function would
decide this only based on the *BUILD* setting, which was supposed to
contain the package under test, but in reality a Maintenance Update can
include multiple packages and the one which is used in the *BUILD*
setting is actually only the first package from the MU when ordered
alphabetically. This of course causes some tests to be skipped even when
it is necessary that they run.

This commit modifies the `lib/hacluster::is_not_maintenance` function to
also query SMELT for the list of packages associated to the Maintenance
Incident, and then check against that list too after having checked
against the value of the *BUILD* setting. This way, the function is more
exhaustive in its checks, and should not skip necessary tests based on
incomplete information. It also adds a check for `kernel`, as kernel
maintenances should not skip any tests.