Reboot SUT if zypper patch in post_migration returns 102
The `migration/online_migration/post_migration` test module includes a
call to `lib/utils::fully_patch_system` as its last task. This call -
which performs a `zypper patch` in the SUT - can return a retval of 102,
which is a suggestion from `zypper` to restart the SUT as newly
installed packages in the SUT require a reboot to start working (for
example, new kernel or new kernel modules). This commit adds a check for
retval 102 after `fully_patch_system` finishes, and if the return value
matches 102, will restart the SUT and reconnect the consoles.