LTP: Set default LTP_TAINT_EXPECTED based on distro
The default value for LTP_TAINT_EXPECTED was for SLE. Therefore for
Tumbleweed it needed to be set in the testsuite setup
(https://openqa.opensuse.org/admin/test_suites).

Having a meaningful value is better for documentation purposes
and is future prove in case we start testing Leap with LTP
(Leap shares SLE kernel, thus variable setup in the testsuite would not
work).

Signed-off-by: Petr Vorel <pvorel@suse.cz>