Record the package list before and after zypper path
Sometimes we need to know the packages changes during our test when we
debugging a failure, especially for QAM test.
Ensure that all of our tests, whether they run on prem or in the cloud,
record the list of packages and their versions before and after the zypper patch command.

Related: https://jira.suse.com/browse/TEAM-7106