Split the qesap regression test module (#20156)
* Fix tag name in vpc_id AWS query

In qe-sap-deployment v0.31.0, the tag name aplied to VPC has been
changed from Workspace to workspace.
Adapt the test code to this change, improve robustness.

* Split the qesap regression test module

Split the module about stesting the deployment in 4 modules each on eto
focus on a different test area. Introduce a new test step to trigger a
Crash in a node.
Move postfail for qesap regression modules to a lib function as CI/CD
blame the change otherwise due to code repetition.
Add related unit tests.