Delay the crash system command
One of the HanaSR test is about crashing one cluster node running HANA.
Crash command is executed through a ssh channel. Problem is that, as soon as the
system crash, the ssh connection  is interrupted leaving the ssh client blocked.
The idea is: compose the remotely executed command with a sleep and then the crash,
run these two in background. It gives time to the ssh client to close the
session before the crash happening.
Remove the timepout=0 behavior, stop forwarding to run_ssh_command all
the args content.