Change error handling for venv related methods
Remove the failok argument and manage everything with return code and by
never die in the low level function. To do so, also pipe to log has to
be managed in a different way to allow both pipefail and pipe to tee to
be properly composed in the command and exit code to be properly
returned to the caller.