Change crash command timeout to a positive value (#20244)
Timeout used as argument for run_ssh_command when sending the crash
command was 0. Zero has a pecial meaning and handling in both
run_ssh_command and below script_run API.
Change it to 10, as thanks to & within the command anf -fn in ssh,
this command should immediately return the prompt.