kernel: Add support for livepatching on SL Micro 6.0+
This change introduces support for SLE Micro 6.0 and newer. There is a
difference in the live patching setup compared to previous
versions. Unlike SLE or SLE Micro 5.x, there is no additional live
patching product. Live patches are now part of the main repository.

The testing logic has also changed. There are no maintenance
incidents, as updates are tested in stagings. The staging repository
is represented by the variable `OS_TEST_REPOS`, which is the same for
other staging tests and is not specific to live patching. Kernel and
live patches tested in staging repositories contain `stage` in their
names and have a `99999` version.