Ignore crmsh deprecation messages when migrating from 12-SP5
During SLES+HA migration tests from 12 to 15, test code needs to deal
with any `clvmd` resources configured in the cluster, as `clvmd` was
available in SLES 12, but not in SLES 15. This is handled by the
`ha/migrate_clvmd_to_lvmlockd` test module, which will check for `clvmd`
resources in the cluster, and if they are present, will perform the
necessary steps to remove the `clvmd` resource and configure in its
place a `lvmlockd` resource, applying also the required changes to the
LVM configuration.

Starting with the `crmsh` version available in 15-SP7, the commands to
remove `clvmd` are failing as now `crmsh` reports more cases of
deprecated resources and deprecated attributes syntax; it also requires
a confirmation due to these warnings. The request for confirmation is
making the test module `ha/migrate_clvmd_to_lvmlockd` fail, so this
commit adds `--force` to the failing `crmsh` commands to ignore the
warnings.

Even though normally it would not be safe to ignore `crmsh` warnings,
in this case since it's a migration scenario from SLES 12-SP5 to 15-SP7,
we must assume cluster configurations under SLES 12 will have many
deprecated resources and attributes syntax, but even with these, the
cluster should come back online after a migration, so we're choosing to
ignore the warnings only in this module which deals only with migrations
from SLES 12.