Run split-brain-iptables cluster crash test with enter_cmd
Starting on 15-SP7, the `crmsh` cluster crash test command
`split-brain-iptables`, causes the crash immediately, preventing
`script_run` from collecting the return value of the command, and
causing a failure in the test module. The other crash commands such as
`kill-sbd` or `kill-corosync` still cause the crash, but crash is
delayed allowing `script_run` to collect the return value of the call.

This commit replaces `script_run` with `enter_cmd` when running
`crm cluster crash_test --split-brain-iptables --force` to avoid
attempting to check for the return value. It also adds a new error
message with `record_info` if `split-brain-iptables` check was run and
it did not cause the fence.