Ensure we run $self->wait_for_cluster in a cluster node
Commit 188213bce43c679782a252ed31cce9a32ac52832 introduced a cluster
check in `tests/sles4sap/publiccloud/qesap_prevalidate` to make sure we
have a healthy cluster right after the deployment. It does this by
calling `lib/sles4sap_publiccloud::wait_for_cluster()`, which relies on
the cloud instance where the cluster is checked to be stored in
`$self->{my_instance}`.

However, this was added after a loop that cycles through all the
deployed cloud instances, and in the case of SBD, this includes the
iSCSI server which is not a member of the cluster. Due to the way Perl
orders lists within hashes, it can happen that the iSCSI server is the
last entry in this list, leaving it stored in `$self->{my_instance}` and
causing `$self->wait_for_cluster()` to fail as the commands are not run
in a member of the cluster.

This commits works around the issue by cycling through the list of instances
again to ensure a cluster node is set in `$self->{my_instance}`. It also
includes a check for the *HA_CLUSTER* setting as it does not make sense
to check for cluster health on non-cluster scenarios.