Add setting to force install from MIRROR_HTTP on PXE boot
This commit introduces the setting *FORCE_INSTALL_FROM_MIRROR_HTTP*,
which when set to true forces `boot/boot_from_pxe` to install the SUT
from the URL contained in *MIRROR_HTTP* (grub boot `install=` argument)
even when running in workers with a closer mirror.

Idea behind this is to allow workers to specify a combination of
*PXE_PRODUCT_NAME* and *MIRROR_HTTP* settings which will lead to a
proper installation of a product/build under test.