Fix use of unitialized value while splitting UPDATE_ADD_CONFLICT
We get a "Use of uninitialized value in split" warning in while splitting the
UPDATE_ADD_CONFLICT variable as this is most often not set. This is fixed by
splitting only when the variable is defined.

Additionally the loop was removed as `push` gets lists as arguments and `split`
returns the list in list context.

Related Ticket: poo#170005