Move wait_still_screen to the start of the loop
There is a loop in the `sles4sap/wizard_hana_install` test module, which
cycles until the `sap-product-installation` needle matches. On SUTs with
only one disk, this is a good solution as it allows the test to continue
faster, but on SUTs with more disks, it's possible that the dialog
prompting the user to select the disk where HANA is going to be
installed takes a few seconds to appear, which can cause the module to
wrongly match with `sap-product-installation` which is the window behind
the disk selection dialog.

This commit moves the `wait_still_screen` call from the end of the loop
to the start of the same loop. Idea is that module will wait until
screen settles before attempting to match either the disk selection
dialog or the `sap-product-installation` needle, giving it more time to
wait for the disk selection dialog to be shown.