Enable docker testing in sle12sp5 LTSS in aarch64
The container module for aarch64 12SP5 is not going to be enabled hence
the container packages will land in LTSS. We need to skip enabling
container repo.

The PR is skipping `validate_btrfs.pm` as the module is dependent on
`/var` subvolume that is not present in aarch64 sle12sp5 installations

Package `zypper-docker` is not yet present and it is unclear whether it
will be added in the near future.

```
devtmpfs        974M     0  974M   0% /dev
tmpfs           992M     0  992M   0% /dev/shm
tmpfs           992M   23M  970M   3% /run
tmpfs           992M     0  992M   0% /sys/fs/cgroup
/dev/vdb2        38G  2.9G   34G   8% /
/dev/vdb2        38G  2.9G   34G   8% /.snapshots
/dev/vdb1       305M  216K  305M   1% /boot/efi
tmpfs           199M     0  199M   0% /run/user/0
```

- ticket: [[Containers] Prepare docker tests on 12-SP5 arm](https://progress.opensuse.org/issues/170251)