Add workaround for bsc#1217583
A specific registercloudguest on a specific SP version has issue that
result in a warning about IPv6 to result in a registration failure due
to an error in its internal exception handling.
Issue has been already fixed but on BYOS images needs image to be
registered that is not possible due to the bug itself.
Temporary use SUSEConnect in this particular situation waiting for an
image refresh.