Add variable for the AWS IBSm project (#21000)
To establish the network peering the the IBSm in AWS, the value of
a tag Project is needed. Value is the one used during the IBSm
deployment. Create a variable to be able to provide it from the
external. Variable is named IBSM_PRJ_TAG. Temporary provide a valid
default value for it. As soon as all JobGroup will have the variable it
needs to be changed to get_required_var without default.