Add a test for cluster_md with gfs2
This commit allows the `ha/filesystem` test module to format the
filesystem for the cluster_md resource test as either an ocfs2 filesystem
or a gfs2 filesystem. Support for gfs2 has ben added in the later
versions of SLES 15, so this module has been configured to:

- Select ocfs2 by default on SLES 15
- Select gfs2 by default on SLES 16
- Allow the default selection to be overwritten by the setting
  **HA_CLUSTER_MD_FS_TYPE**

Changes apply only when configuring a cloned filesystem (a filesystem
which is active in all nodes of the cluster). Primary/secondary
filesystem resources such as drbd will still use xfs.