Add unattended installation using iscsi ibft disk
Add unattended installation using iscsi ibft disk for agama. Based on
bsc#1239046, the ibft is not supported on SLE16 yet, so the ibft
configuration can not be found on SLE16, while SLE16 is using Agama
installer, so change the search log to agama general log. We will update
the log path if necessary after the feature is implemented.