Remove IPXE_HTTPSERVER from test schedule to support more complex setups
This variable depends on the worker and its according network structure
and should be set as a worker setting.
https://gitlab.suse.de/openqa/salt-pillars-openqa/-/merge_requests/970
is an example how this should be set. For o3 and other infrastructure
this chance should make no difference as the previous value was not
accessible outside of SUSE.

Related ticket: https://progress.opensuse.org/issues/174298