Use root-console for php_postgresql case on aarch64
On aarch64, sometimes test will fail due to serial console stuck, using
`root-console` for part of test due to:
1. when the sessoin stuck, we couldn't restart serial-getty service and
reconnect to it, because the `pushd` `popd` records will lost which
cause failure.
2. If the whole test module uses `root-console`, the module running time
will increase to more than 6m from 2m. So use it for a part of the
module.

Related: https://progress.opensuse.org/issues/178639