Add test module to ensure HA is activated in SLES 16
It is not possible yet to activate the HA Extension in current builds of
SLES 16 during installation. This commit adds a new test module included
in the `agama_install_sles_ha.yaml` schedule which will verify that the
installed system is:

- Registered to SCC
- Using the SCC proxy
- /etc/SUSEConnect exists
- HA Extension is enabled

In case any of the above conditions is not met, the module will register
the SUT to SCC using the proxy and activate the HA Extension, while soft
failing on known bugs and tickets related to the failed conditions.

Result should be a SLES+HA 16 system with the HA Extension activated and
the `ha_sles` pattern installed.