Copy agama jsonnets for SLES for SAP to data/sles4sap/agama
The jsonnet files under `data/yam/agama/auto` are expected to be under
constant modification and recommendation from YAM squad is to instead
use the files there as a template for dedicated copies to be used by
other squads. This commit creates/copies the jsonnet files for SLES for
SAP on qemu and pvm_hmc.