Add SAP Business one module on SLE & HA
The SAP Business One module contains the specialized tools for the installation
of SAP Business One product. The module is maintained and supported by the
SUSE Linux Enterprise Server product subscription.

Is needed for installation of bone-installation-wizard on SLE or HA system
nothing provides 'patterns-sap-bone' needed by the to be installed
bone-installation-wizard