Add alternative to cs_wait_for_idle when not present
The package `ClusterTools2` which includes commands such as
`cs_wait_for_idle` to wait until a cluster is in idle state before
applying more changes, is not always available.

In SLES 12 and 15, the package is part of the SAP module, and as such it's
not available in a SLES system with HA Extension activated. It is
available in SLES for SAP 12 and 15 however.

In SLES 16, it is available in the HA Extension.

Since test modules cannot know in advance whether a SUT will have
`ClusterTools2` installed, the function
`hacluster::wait_for_idle_cluster` would check if `ClusterTools2` is
installed, and call `cs_wait_for_idle`; when not installed, it would not
wait and simply return. This can lead to issues, specially on tests
running in the serial terminal, which is faster than the root console.

This commit implements an alternative waiting code for systems where
`ClusterTools2` is not available. It uses the command `crmadmin` (part
of `crmsh`) which should be available always, and which is one of the
commands used by `cs_wait_for_idle`.

Also included in the commit are changes in the `ha/filesystem` test
module to use `hacluster::wait_for_idle_cluster` instead of `sleep`
while waiting for an idle cluster, and the move of this module to the
serial terminal when possible, as it's faster and less susceptible
to typing errors than the console.

Also included in the commit are changes in the `ha/haproxy` and
`sles4sap/ensa/ensa2_web_methods_checks` test modules so they call
`hacluster::wait_for_idle_cluster` after changing the configuration
of the cluster.