Fix iSCSI devices persistence on node reboot
The module `ha/iscsi_client_setup` which adds remote iSCSI devices to
the System Under Test, is not configuring the devices to be available on
a reboot (see:
http://mango.qe.nue2.suse.org/tests/6470#step/check_after_reboot/17).

This commit modifies the setup commands to configure the iSCSI devices
so they are configured on reboots.

Also, the commit moves the module so it runs on the serial terminal. The
serial terminal is faster and less susceptible to typing errors, and
since differently than the `ha/iscsi_client` test module which uses a
`yast` package, the `ha/iscsi_client_setup` uses only the command line
so it works also on the serial terminal.