Select root-console without assert and clean
Some HA modules can cause a fence operation, and they need to run
in the root console, so these modules or the ones that follow can
properly assert the grub2 menu or the boot menu in the console.
This applies specifically to:

- ha/fencing
- ha/priority_fencing_delay
- ha/qnetd
- ha/check_hawk
- ha/ha_cluster_crash_test

However, when running HA tests in the serial terminal, the root
console can be filled by messages which obscure the root prompt,
causing the test to fail on a needle.

This commit introduces a new function in `lib/hacluster` to select
the console first with `await_console => 0` to skip the needle check,
and then clearing it and sending a `ret` key to force it to show the
prompt, before asserting the `root-console` needle.