Close terminal session and reset consoles in ha/await_upgrade_or_update
The module `migration/version_switch_upgrade_target` includes a call to
`reset_consoles` when scheduled in non-PVM backends. In HA tests, it is
usually followed by either `ha/cluster_state_mgmt` or
`ha/cluster_boot_mgmt`; both of these, will run `select_serial_terminal`
as their first step, but after a `reset_consoles`, this function will
look for a `login:` prompt in the terminal. As there is nothing closing
the session in the terminal, this can lead to failures.

As a workaround, this commit closes the session in the serial terminal,
as well as resets the consoles to allow the `select_serial_terminal`
calls that follow to find a `login:` prompt and continue.