Improve sles4sap::check_instance_state
The `check_instance_state` method in `lib/sles4sap` wrongly returns
successfully when the function `GetProcessList` does not list any
process. This is because it was only checking for lines not matching any
of the expected status texts (GREEN or GRAY) in the output and ignoring
any other lines.

This commit adds a check to verify that `GetProcessList` actually
printed the processes in the output.

Also included is a small modification so the module
`sles4sap/netweaver_patch` runs in the root console; and the module
`sles4sap/sap_suse_cluster_connector` runs in the root console in SLES
for SAP 12.