Improve sles4sap::check_instance_state
The `check_instance_state` method in `sles4sap` package wrongly returns
successfully when the function `GetProcessList` does not list any
process. This is because it was only checking for lines not matching any
of the expected status texts (GREEN or GRAY) in the output and ignoring
any other lines.

This commit adds a check to verify that `GetProcessList` actually
printed the processes in the output.