Allow HanaSR test to support both peering methods
Add some code to allow supporting both the methods to create the network
peering to the IBSm in Azure: using the az cli (legacy) and using
terraform (new). The problem is to be able to support both with the same
conf.yaml. This commit is to avoid failure in qesap_terraform about
missing IBSM_VNET and IBSM_RG openQA variables.