Adjust password length according to vnc RFC
The 8 char password max is a protocol limitation, rather than an implementation issue.
See [1]:

"The client encrypts the challenge with DES, using a password supplied by the user as
the key. To form the key, the password is truncated to eight characters, or padded
with null bytes on the right."

[1] https://www.rfc-editor.org/rfc/rfc6143#section-7.2.2