Download script when openqa-bootstrap runs in Leap
Use always the latest bootstrap script from the repo as the rpm is not
always the latest one. This will fix
https://progress.opensuse.org/issues/180905 where the openQA config files have
been moved around but the bootstrap script is not up to date. As for now there
is not automatic submission on Leap and this is the easiest approach to fix
the test module.